REFACTOR: Centralised stream response types
